+++ tests/pica/mod.rs
@@ -9,7 +9,6 @@ mod partition;
 mod print;
 mod sample;
 mod select;
-mod slice;
 mod split;
 mod xml;