+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * #572 Rephrase misleading documentation (`invalid`)
+* #590 Stabilize `slice` command
 
 ## 0.14.1